Restructured the project a bit
