+++ Apps/gm-flutter-app/gimmillions/lib/models/stock-data.dart
