--- Apps/gm-flutter-app/gimmillions/lib/views/stocks/futures.dart
@@ -1,13 +1,0 @@
-import 'package:flutter/material.dart';
-
-class FuturesWidget extends StatefulWidget {
-  @override
-  _FuturesState createState() => _FuturesState();
-}
-
-class _FuturesState extends State<FuturesWidget> {
-  @override
-  Widget build(BuildContext context) {
-    return Container();
-  }
-}
